+++ source/extensions/lightspeed.trex.mesh_properties.shared.widget/lightspeed/trex/mesh_properties/shared/widget/tests/e2e/test_widget.py
@@ -1135,12 +1135,14 @@ class TestSelectionTreeWidget(AsyncTestCase):
         await ui_test.human_delay(3)
 
         # Set a random remix category for us to assign
-        box = ui_test.find("Remix Categories//Frame/**/CheckBox[*].name=='remix_category:world_ui'")
+        box = ui_test.find("Add Remix Categories to Prim//Frame/**/CheckBox[*].name=='remix_category:world_ui'")
         await ui_test.emulate_mouse_move_and_click(box.position)
         await ui_test.human_delay(3)
 
         # Assign the category
-        assign_button = ui_test.find("Remix Categories//Frame/**/Button[*].identifier=='AssignCategoryButton'")
+        assign_button = ui_test.find(
+            "Add Remix Categories to Prim//Frame/**/Button[*].identifier=='AssignCategoryButton'"
+        )
         await assign_button.click()
         await ui_test.human_delay(3)
 
@@ -1173,15 +1175,17 @@ class TestSelectionTreeWidget(AsyncTestCase):
         await ui_test.human_delay(3)
 
         # Set a random remix category for us to assign
-        box = ui_test.find("Remix Categories//Frame/**/CheckBox[*].name=='remix_category:world_ui'")
+        box = ui_test.find("Add Remix Categories to Prim//Frame/**/CheckBox[*].name=='remix_category:world_ui'")
         await ui_test.emulate_mouse_move_and_click(box.position)
         await ui_test.human_delay(3)
-        box2 = ui_test.find("Remix Categories//Frame/**/CheckBox[*].name=='remix_category:decal_Static'")
+        box2 = ui_test.find("Add Remix Categories to Prim//Frame/**/CheckBox[*].name=='remix_category:decal_Static'")
         await ui_test.emulate_mouse_move_and_click(box2.position)
         await ui_test.human_delay(3)
 
         # Assign the category
-        assign_button = ui_test.find("Remix Categories//Frame/**/Button[*].identifier=='AssignCategoryButton'")
+        assign_button = ui_test.find(
+            "Add Remix Categories to Prim//Frame/**/Button[*].identifier=='AssignCategoryButton'"
+        )
         await assign_button.click()
 
         await ui_test.human_delay(3)