+++ source/extensions/lightspeed.common/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.6]
+### Changed
+- Updated Remix Categories list
+
 ## [1.0.5] - 2024-10-18
 ### Added
 - Added constants for "focus in viewport" widget tooltips