+++ source/extensions/lightspeed.trex.contexts/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.6]
+### Added
+- Updating dialog removal calls
+
 ## [1.0.5] - 2024-06-26
 ### Changed
 - Added an argument to the global event call