+++ source/extensions/lightspeed.trex.mesh_properties.shared.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.9]
+### Changed
+- Restructured categories window and added more documentation for users
+
 ## [1.1.8]
 ## Fixed
 - Fixed separators and spacing layout